data structures for detailed latency histograms
No functional change, only data structures.
Courtesy of Tony Finch @ Internet Systems Consortium.

The size of a hg64 histogram depends on the range of values in the data
stream, not the length of the data stream. The update performance is
roughly constant, and mostly depends on cache locality.

Given our small range of values (seconds) and low precision required,
it takes just couple hundreds of bytes per thread.