add detailed latency histograms
- `dnsperf`:
  - New option `-O latency-histogram=true` prints detailed table of
    latencies every time statistics are printed. This includes
    statistics printed by `-S` option when `-O
    verbose-interval-stats=true` is in use.

Worst case observed in benchmarks:
On a totally maxed-out 16-core system doing 1.5 M QPS, where dnsperf is
the bottleneck, the impact is around 1.5 % QPS drop compared to the
previous version. Tested between two AWS c5n.4xlarge instances, dnsperf
on one end and DNS echo server on the other end.

In sane scenarios (when dnsperf is not the bottleneck) the impact negligible.